pyproject.toml
[tool.poetry]
name = "doedeer-gendiff"
version = "1.0.0"
description = "Show differnces between two files"
authors = ["Doe Deer <gregorylos97@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.6"
PyYaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
coverage = "^5.2.1"
wemake-python-styleguide = "^0.14.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.generate_diff:main"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

